requirements.txt
paho-mqtt
redis
pytest
psycopg2
pandas
rrcf
librosa